Split options for Marching and RegularSpace into hpp/cpp.
