+++ components/wmtk_components/marching/CMakeLists.txt
@@ -1,6 +1,7 @@
 
 set(SRC_FILES
     internal/MarchingOptions.hpp
+    internal/MarchingOptions.cpp
     internal/Marching.hpp
     internal/Marching.cpp
 	marching.hpp
